Refactor package into layered modules
