+++ src/dc43/components/data_quality/manager.py
@@ -4,7 +4,7 @@ from __future__ import annotations
 
 from open_data_contract_standard.model import OpenDataContractStandard  # type: ignore
 
-from dc43.services.data_quality.models import ObservationPayload
+from dc43.lib.data_quality import ObservationPayload
 
 from .engine import ValidationResult, evaluate_observations
