+++ src/dc43/services/data_quality/__init__.py
@@ -1,7 +1,13 @@
-"""Client utilities for interacting with data-quality services."""
+"""Client utilities for interacting with data-quality services.
+
+Prefer importing :class:`ObservationPayload` from :mod:`dc43.lib` when writing
+pure library integrations. The symbol remains available here for backwards
+compatibility with existing integrations.
+"""
+
+from dc43.lib.data_quality import ObservationPayload
 
 from .client import DataQualityServiceClient, LocalDataQualityServiceClient
-from .models import ObservationPayload
 
 __all__ = [
     "DataQualityServiceClient",
