+++ src/dc43/services/governance/client.py
@@ -8,7 +8,7 @@ from open_data_contract_standard.model import OpenDataContractStandard  # type: 
 
 from dc43.components.data_quality.engine import ValidationResult
 from dc43.components.data_quality.governance import DQStatus
-from dc43.services.data_quality.models import ObservationPayload
+from dc43.lib.data_quality import ObservationPayload
 
 from .models import (
     GovernanceCredentials,
